requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aioresponses==0.7.8
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
asynctest==0.13.0
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.7
django-celery-beat==2.7.0
django-jazzmin==3.0.1
django-timezone-field==7.1
djangorestframework==3.15.2
flower==2.0.1
frozenlist==1.5.0
humanize==4.12.1
idna==3.10
iniconfig==2.1.0
kombu==5.5.0
multidict==6.2.0
packaging==24.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
pytest==8.3.5
pytest-aiohttp==1.1.0
pytest-asyncio==0.26.0
pytest-django==4.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
